pyproject.toml
[project]
name = "muellabfuhr-reminder-python"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "icalendar>=6.0",
    "paho-mqtt>=2.0",
    "python-dotenv>=1.2.2",
    "requests>=2.32",
]

[dependency-groups]
dev = ["pytest>=9.0.3"]

[tool.pytest.ini_options]
pythonpath = ["."]
